requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pydantic
python-multipart
openai
pdfminer.six
pillow
pytesseract
python-dotenv
aiofiles
requests